greybark-intelligence/requirements.txt
feedparser>=6.0
telethon>=1.36
pyyaml>=6.0
python-dateutil>=2.8
python-dotenv>=1.0
